### 2024-08-09 #7095 @dfm
In the meantime, please add a section called "Statement of need" to your paper and then comment `@editorialbot generate pdf` on this thread.
