### 2024-08-19 #7118 @editorialbot
[REVIEW]: seismolab: A Python package for analyzing space-based observations of variable stars
**Submitting author:** @astrobatty (<a href="http://orcid.org/0000-0002-8585-4544">Attila Bódi</a>)
**Repository:** https://github.com/konkolyseismolab/seismolab
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.4
