### 2024-09-09 konkolyseismolab/seismolab#3 @darthoctopus
documentation: statement of need
First, I'd like to say that the documentation you've put together is extremely impressive, from a technical perspective! Only one issue remains per the JOSS checklist:

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

I do not believe this is explicitly addressed yet. Perhaps this can go in the "About seismolab" page, which is pretty empty at the moment except for funding information?

ref: https://github.com/openjournals/joss-reviews/issues/7118#issuecomment-2302742402
