### 2024-09-09 konkolyseismolab/seismolab#4 @darthoctopus
paper: statement of need
quoting again from the review criteria:

> A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?

I believe that the statement of need in the current paper only addresses the first of these three (i.e. describing the six main areas of functionality provided by the package). Some description of 1. the intended target audience for seismolab, and 2. the role of seismolab given the present state of the field, would probably be needed to satisfy these other criteria.

ref: https://github.com/openjournals/joss-reviews/issues/7118#issuecomment-2302742402
