### 2024-09-09 konkolyseismolab/seismolab#5 @darthoctopus
paper: state of the field
Per the review criteria:

> State of the field: Do the authors describe how this software compares to other commonly-used packages?

I do not believe that this has been described in the paper yet. Most of the contextualisation provided there is regarding the scientific context, rather than pertaining to methodology/implementations.

ref: https://github.com/openjournals/joss-reviews/issues/7118#issuecomment-2302742402
