### 2024-10-03 #7118 @warrickball
Thanks @keatonb for your commitment!

@astrobatty, I'm not sure you've noticed that @darthoctopus opened three issues directly in the software repo. This is a perfectly reasonable way for the review to proceed but you may want to check that you aren't missing notifications about that activity.
