### 2024-10-05 #7118 @keatonb
I was unable to install seismolab on my machine following the installation instructions. Please see issue https://github.com/konkolyseismolab/seismolab/issues/6 and advise.
