### 2024-10-05 konkolyseismolab/seismolab#6 @warrickball
I can reproduce this with Python 3.12.6 on Fedora 40. I don't run into this issue with Python 3.9.5 on RHEL 8.8 so I suspect it's a Python version issue.

In particular, the Theano dependency is coming from PyMC3. It looks to me like PyMC3 has had little to no support for [about two years](https://www.pymc.io/blog/v4_announcement.html), following what looks like Theano development ending. PyMC is now up to major version 5 (there's some history [in the PyMC blog](https://www.pymc.io/blog/PyMC_Past_Present_Future.html)), so unless there's good reason to persist with older versions of PyMC, I'd suggest trying to upgrade `seismolab` to use that instead.

FWIW I just installed PyMC 5.12.0 on Python 3.9.5 and PyMC 5.17.0 on Python 3.12.6.
