### 2024-10-07 konkolyseismolab/seismolab#6 @darthoctopus
Chiming in with success on different python versions: I've been able to install the package on python 3.10.14 and 3.11.8 (both on Arch Linux)
