### 2024-10-07 26f3172
TemplateFitting: update MCMC sampler from pymc3 to pymc
