### 2024-10-07 8028bbc
Python must be >=3.10 due to issues with arviz
