### 2024-10-07 2d4f7c8
Add python 3.11 & 3.12 to GitHub Actions
