### 2024-10-07 konkolyseismolab/seismolab#6 @astrobatty
Thanks @keatonb for spotting this issue and @warrickball for the useful suggestion. I updated seismolab to use PyMC and added python 3.12 to GitHub Actions.
