### 2024-10-07 22e399d
TemplateFitting: use Uniform priors instead if Normal ones
