### 2024-10-07 da2b671
TemplateFitting: BugFix: use the raw input error values in the likelihood
