### 2024-10-07 0b86b05
TemplateFitting: use smaller font size in arviz pair plot
