### 2024-10-07 9a16b3d
Updated documentation in "about" section
