+++ paper/paper.md
@@ -23,11 +23,13 @@ Classical pulsating stars are characterized by periodic or multi-periodic bright
 
 Upon entering the era of photometric space missions, the launch of the NASA *Kepler* and *TESS* missions have brought a new opportunity to expand the science of variable star, enabling the characterization of short-term variability and the detection of millimagnitude-level variations. The latest generation of large photometric and astrometric surveys has greatly expanded the number of known, observed and classified variable stars. Mining these large data sets has led to new discoveries and more detailed analysis in this field [@Plachy21, @Molnar22, @Benko23].
 
-Several techniques have been developed to search for periodicities of the light curves and also for any deviation from the strictly periodic behavior. We have developed a Python package, ``seismolab``, which implements various methods for downloading, analyzing, and visualizing data of variable stars from space-based surveys. The framework is primary intended to be used with data obtained by the *Kepler*, *TESS* and *Gaia* surveys, but can also be used by other similar existing and future surveys. Some modules are also useful for analyzing ground-based observations.
+Several techniques have been developed to search for periodicity in light curves and also for any deviation from the strictly periodic behavior. We have developed a Python package, ``seismolab``, which implements various methods for downloading, analyzing, and visualizing data of variable stars from space-based surveys. The framework is primary intended to be used with data obtained by the *Kepler*, *TESS* and *Gaia* surveys, but can also be used by other similar existing and future surveys. Some modules are also useful for analyzing ground-based observations.
 
 # Statement of need
 
-The purpose of ``seismolab`` is sixfold. It is able to combine *Gaia* data with the Bailer-Jones distance catalog [@BJ21], galactic extinction maps [@Bovy16] and magnitudes from the Simbad catalog [@Wenger00]. Different modules are implemented to extract the Fourier coefficients and Fourier parameters of light curves, and to derive the temporal variation of the amplitude, phase and zero point of the dominant variation [@Benko23]. One can extract the minimum or maximum times and derive an O-C diagram [@Sterken05]. The package can be use to fill gaps in time series data using the method of inpainting [@Pires09, @Pires15]. There is also a module that provides different transformation methods for time-frequency analysis [@Kollath97].
+``seismolab`` is a fully-python based, open-source package, built on top of popular Python packages such as Numpy (Harris, C.R. et al., 2020, Nature, 585, 357–362), Scipy (Virtanen, P. et al., 2020, Nature Methods, 17(3), 261-272), PyMC (Abril-Pla, O. et al. (2023), PEERJ-CS, 9(15), 1516) and Astropy (Astropy Collaboration et al., 2022, ApJ, 935, 167). The framework contains the main analysis tools for the astronomical community working on the light curves of variable stars, primarily, but not exclusively, from space-based surveys. ``seismolab`` is a modular library that allows users to select the best method(s) for their particular science problem. It has six main modules which implement common operations often used by the variable star community, but available in limited or no form in other popular codes.
+
[…]
