### 2024-10-07 konkolyseismolab/seismolab#6 @warrickball
I just successfully installed `seismolab` 1.1.4 on Python 3.12.6 / Fedora 40.

I'm also going to mention openjournals/joss-reviews#7118 to cross-reference this back to the JOSS review.
