### 2024-10-07 konkolyseismolab/seismolab#3 @astrobatty
>First, I'd like to say that the documentation you've put together is extremely impressive, from a technical perspective!

Thanks @darthoctopus !

I updated the documentation accordingly.
