### 2024-10-07 konkolyseismolab/seismolab#4 @astrobatty
I updated the paper.
