### 2024-10-07 konkolyseismolab/seismolab#5 @astrobatty
I updated the paper.
