### 2024-10-15 #7118 @warrickball
Hi @keatonb & @darthoctopus, I was wondering how your reviews were proceeding? @astrobatty has fixed the installation issue @keatonb was having (AFAICT) and modified the paper according to the issues that @darthoctopus openend.
