### 2024-10-16 konkolyseismolab/seismolab#6 @keatonb
I have also been able to install the package now. Thanks for working on this. I will close this issue.
