### 2024-10-16 konkolyseismolab/seismolab#7 @keatonb
Error running example gaia query
Hi, I'm following the tutorial example for querying Gaia and I encounter an error. I am using a Mac with macOS 14.6 (Sonoma) and an Apple M3 processor, using Python v3.12.4.

```python
from seismolab.gaia import query_gaia

targets = [1022096392749670016,3107142829863485952]

result_table = query_gaia(targets)
```

Output is:
```python
Using Combined19 map from mwdust
Querying Gaia DR3 and BJ catalogs...
DALQueryError                             Traceback (most recent call last)
Cell In[2], line 3
      1 targets = [1022096392749670016,3107142829863485952]
----> 3 result_table = query_gaia(targets)

File ~/anaconda3/lib/python3.12/site-packages/seismolab/gaia/query_gaia.py:142, in query_gaia(targets, gaiaDR, use_photodist, dustmodel, plx_offset)
    140 # --- Query Gaia -----------
    141 try:
--> 142     gaiaquery, BJdist = perform_query(targets,useEDR3)
    143 except requests.exceptions.ConnectionError:
    144     sleep(1)

File ~/anaconda3/lib/python3.12/site-packages/seismolab/gaia/querytools.py:369, in perform_query(data, useEDR3)
    366 service = vo.dal.TAPService("https://dc.zah.uni-heidelberg.de/tap")
    368 # Query BJ catalog
--> 369 BJdist = service.search("SELECT TOP 1000000 BJ.source_id, \
    370                     BJ.r_med_geo,BJ.r_lo_geo,BJ.r_hi_geo, \
    371                     BJ.r_med_photogeo,BJ.r_lo_photogeo,BJ.r_hi_photogeo \
    372                     FROM gedr3dist.litewithdist as BJ \
    373                     WHERE BJ.source_id "+qrelation+" "+targetlist)
    374 BJdist = BJdist.to_table()
    376 # Query Gaia catalog

File ~/anaconda3/lib/python3.12/site-packages/pyvo/dal/tap.py:278, in TAPService.run_sync(self, query, language, maxrec, uploads, **keywords)
    249 def run_sync(
    250         self, query, language="ADQL", maxrec=None, uploads=None,
    251         **keywords):
    252     """
    253     runs sync query and returns its result
    254
   (...)
    274     TAPResults
    275     """
    276     return self.create_query(
    277         query, language=language, maxrec=maxrec, uploads=uploads,
--> 278         **keywords).execute()

File ~/anaconda3/lib/python3.12/site-packages/pyvo/dal/tap.py:1117, in TAPQuery.execute(self)
   1103 def execute(self):
   1104     """
   1105     submit the query and return the results as a TAPResults instance
   1106
   (...)
   1115        for errors parsing the VOTable response
   1116     """
-> 1117     return TAPResults(self.execute_votable(), url=self.queryurl, session=self._session)

File ~/anaconda3/lib/python3.12/site-packages/pyvo/dal/adhoc.py:111, in AdhocServiceResultsMixin.__init__(self, votable, url, session)
    110 def __init__(self, votable, url=None, session=None):
--> 111     super().__init__(votable, url=url, session=session)
    112     self._adhocservices = list(
    113         resource for resource in votable.resources
    114         if resource.type == "meta" and resource.utype == "adhoc:service"
    115     )

File ~/anaconda3/lib/python3.12/site-packages/pyvo/dal/query.py:322, in DALResults.__init__(self, votable, url, session)
    320 self._status = self._findstatus(votable)
    321 if self._status[0].lower() not in ("ok", "overflow"):
--> 322     raise DALQueryError(self._status[1], self._status[0], url)
    324 if self._status[0].lower() == "overflow":
    325     warn("Partial result set. Potential causes MAXREC, async storage space, etc.",
    326          category=DALOverflowWarning)

DALQueryError: connection pool is closed
```
