### 2024-10-16 #7118 @keatonb
I was able to install the software and I am resuming testing. I encounter an error with the Gaia module https://github.com/konkolyseismolab/seismolab/issues/7. By the way, how should I run all the unit tests in the /tests/ directory?
