### 2024-10-16 c5485fb
Add M1 Mac to GitHub Actions
