+++ .github/workflows/main.yml
@@ -7,7 +7,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest]
+        os: [ubuntu-latest, macos-latest, macos-latest-xlarge]
         python-version: ["3.10", "3.11", "3.12"]
     steps:
     - name: Checkout seismolab
