### 2024-10-16 473f12e
Revert "Add M1 Mac to GitHub Actions"
This reverts commit c5485fb6d110be42b1818ceddad8222710d672ce.
