### 2024-10-16 konkolyseismolab/seismolab#7 @astrobatty
Hmm, that's weird. This seems to be a problem with `pyvo`, not `seismolab`. Unfortunately, I don't have an M3 Mac (yet) and GitHub Actions with M1 Mac architecture isn't free, so I can't test this.

Could you run this snippet?

```python
import pyvo as vo
service = vo.dal.TAPService("https://dc.zah.uni-heidelberg.de/tap")

BJdist = service.search("SELECT TOP 10 BJ.source_id, \
                        BJ.r_med_geo,BJ.r_lo_geo,BJ.r_hi_geo, \
                        BJ.r_med_photogeo,BJ.r_lo_photogeo,BJ.r_hi_photogeo \
                        FROM gedr3dist.litewithdist as BJ")
BJdist = BJdist.to_table()

print(BJdist)
```

The expected output is:
```python
 source_id   r_med_geo  r_lo_geo  r_hi_geo r_med_photogeo r_lo_photogeo r_hi_photogeo
                 pc        pc        pc          pc             pc            pc
------------ --------- --------- --------- -------------- ------------- -------------
  4295806720 2530.1523 1989.7006 3381.6545       2555.457     2213.3289     3038.3936
 34361129088 306.34778 294.81198  316.8109       310.8287     300.25058     319.78366
 38655544960 314.09866 312.24747 316.44965       313.4547     311.59113      315.8233
309238066432 745.71185  589.6301 980.24524      894.88654      740.2594     1037.2533
343597448960  3076.442  2366.115  4843.747      2312.6553     2050.8374       2602.06
515396233856   6746.08 1995.6241  10343.49        5598.27      4505.232      6764.955
549755818112  614.9362 604.02435 626.47345       616.6514     603.03955     628.16156
828929527040  814.8725 780.01373 868.96045      811.62646      768.9585      852.5948
927713095040 2507.6306 1645.8883  4422.298      1666.9054     1457.9774     1936.1094
966367933184 819.98364  590.4627  1262.996       859.0772      731.9902     1032.5143
```
