### 2024-10-17 #7118 @darthoctopus
Update from me: the unit tests run successfully for me, and I have been working my way through some of the example notebooks before I tick the box on functionality. I have unfortunately fallen behind on that; apologies.
