### 2024-10-17 konkolyseismolab/seismolab#3 @darthoctopus
Looks good to me, thanks!
