### 2024-10-17 konkolyseismolab/seismolab#4 @darthoctopus
Looks good to me, thanks!
