### 2024-10-17 konkolyseismolab/seismolab#5 @darthoctopus
Looks good to me, thanks!
