### 2024-10-17 konkolyseismolab/seismolab#7 @keatonb
You know what? I tried it again this evening and it works. Must have been a connection issue with the Gaia server earlier today. Closing this issue.
