### 2024-10-17 #7118 @keatonb
> I was able to install the software and I am resuming testing. I encounter an error with the Gaia module [konkolyseismolab/seismolab#7](https://github.com/konkolyseismolab/seismolab/issues/7). I ran all the unit tests and the only ones that failed were related to the gaia queries.

Works for me now. Must have been a temporary issue with the Gaia server.
