### 2024-11-01 #7118 @warrickball
Thanks @keatonb & @darthoctopus, and congratulations @astrobatty! I'll start the [post-review process](https://joss.readthedocs.io/en/latest/submitting.html#review-process), the main element of which is

> * Upon successful completion of the review, authors will make a tagged release of the software, and deposit a copy of the repository with a data-archiving service such as [Zenodo](https://zenodo.org/) or [figshare](https://figshare.com/), get a DOI for the archive, and update the review issue thread with the version number and DOI.

If you go ahead and deposit an archive with e.g. Zenodo, let me know the URL here and I'll add the metadata to the review.

I'll start doing some editorial checks on the paper.
