### 2024-11-05 #7118 @warrickball
@astrobatty — Great, thanks! For whatever reason, the Zenodo archive is indicating that the archive's licence is CC-BY 4.0. Could you change this to match the repository's MIT Licence? You should be able to change the metadata without creating a new archive.

I'll have a first proofread of the paper now.
