### 2024-11-05 #7118 @warrickball
I've opened a [pull request](https://github.com/konkolyseismolab/seismolab/pull/8) with a first set of editorial changes. Once those are merged (and the archive licence matches the repo) I'll do a test run of the publication process and we can iron out any remaining editorial details.
