### 2024-11-06 #7118 @astrobatty
> @astrobatty — Great, thanks! For whatever reason, the Zenodo archive is indicating that the archive's licence is CC-BY 4.0. Could you change this to match the repository's MIT Licence? You should be able to change the metadata without creating a new archive.

It should be fine now.
