### 2024-11-06 #7118 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
